requirements.txt
pandas==2.3.1
numpy==2.3.1 
openpyxl==3.1.5
dbt-core==1.10.4
dbt-postgres==1.9.0
scikit-learn==1.7.0
matplotlib==3.10.3
seaborn==0.13.2
SQLAlchemy>=1.4
psycopg2-binary>=2.9
